Dodanie edycji histori pojazdu

+++ app/src/main/java/com/example/strzala/e_ksiazaka1/MainActivity.java
@@ -1,10 +1,13 @@
 package com.example.strzala.e_ksiazaka1;
 
+import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.hardware.Camera;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.net.Uri;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
@@ -17,6 +20,11 @@ import android.widget.CheckBox;
 import android.widget.CompoundButton;
 import android.widget.EditText;
 import android.widget.Toast;
+
+import java.io.IOException;
+import java.net.InetSocketAddress;
+import java.net.Socket;
+import java.net.SocketAddress;
 
 import me.dm7.barcodescanner.zxing.ZXingScannerView;
 
@@ -42,29 +50,46 @@ public class MainActivity extends AppCompatActivity {
                 Toast.LENGTH_LONG).show();
     }
 
-    //hash password
+
+
     private void hash()
     {
-      try{
-          hash = "%02320%xwc48" + String.valueOf(password_act.hashCode());
+            if(!password_act.equals("%02320%xwc48")) {
+                try {
+                    hash = "%02320%xwc48" + String.valueOf(password_act.hashCode());
 
 
-      }catch (Exception e)
-      {
-          Log.i("hash",""+e);
-      }
+                } catch (Exception e) {
+                    Log.i("hash", "" + e);
+                }
+            }
+
+
     }
 
 
-    //select data from table user
+    //select data from table user, tworzenie tabel jeśli nie istnieją
     private void SelectDataUser()
     {
 
         try{
-
+            password="";
             sampleDB = this.openOrCreateDatabase(SAMPLE_DB_NAME, MODE_PRIVATE,null);
+
+            //tworzenie tabeli uzytkownik jezeli nie istnieje
             sampleDB.execSQL("CREATE TABLE IF NOT EXISTS uzytkownik (Id INTEGER PRIMARY KEY AUTOINCREMENT,data_dod VARCHAR," +
-                    " email VARCHAR, haslo VARCHAR, zapisz_haslo VARCHAR, qr_code VARCHAR, punkty VARCHAR,admin VARCHAR)");
+                    " email VARCHAR, haslo VARCHAR, zapisz_haslo VARCHAR, qr_code VARCHAR, punkty VARCHAR,admin VARCHAR,czy_zapis VARCHAR)");
+            //tworzenie tabeli samochod jezeli nie istnieje
+            sampleDB.execSQL("CREATE TABLE IF NOT EXISTS samochod (Id INTEGER PRIMARY KEY AUTOINCREMENT,marka VARCHAR," +
+                    "model VARCHAR, rocznik VARCHAR, silnik VARCHAR, nr_rejestracyjny VARCHAR, qr_code VARCHAR,wyswietl VARCHAR)");
+
+            //tworzenie tabeli zgloszenie jezeli nie istnieje
+            sampleDB.execSQL("CREATE TABLE IF NOT EXISTS zgloszenie (Id INTEGER PRIMARY KEY AUTOINCREMENT,zdjecie_przed BLOB," +
+                    "zdjecie_po BLOB, cena_czesci VARCHAR, cena_uslugi VARCHAR, uwagi VARCHAR, data_dod VARCHAR,data_wykonania VARCHAR," +
+                    "status VARCHAR, qr_code VARCHAR, akceptacja VARCHAR)");
+            //tworzenie tabeli naprawy jezeli nie istnieje
+            sampleDB.execSQL("CREATE TABLE IF NOT EXISTS kategorie (Id INTEGER PRIMARY KEY AUTOINCREMENT,nazwa VARCHAR," +
+                    "punkty VARCHAR, kat_1 VARCHAR, kat_2 VARCHAR)");
 
             Cursor c = sampleDB.rawQuery("Select * from uzytkownik where email = '"+login_act+"' ",null);
 
@@ -90,17 +115,23 @@ public class MainActivity extends AppCompatActivity {
 
     private void pamiec_hasla()
     {
-        sampleDB = this.openOrCreateDatabase(SAMPLE_DB_NAME, MODE_PRIVATE,null);
-        sampleDB.execSQL("CREATE TABLE IF NOT EXISTS pamiec (Id INTEGER PRIMARY KEY AUTOINCREMENT," +
-                "email VARCHAR, haslo VARCHAR)");
-        sampleDB.execSQL("Delete from pamiec");
-        sampleDB.execSQL("Insert into pamiec (email,haslo) VALUES ('"+login+"','"+hash+"') ");
-        sampleDB.close();
+        if(login_pamiec.equals("")) {
+            sampleDB = this.openOrCreateDatabase(SAMPLE_DB_NAME, MODE_PRIVATE, null);
+            sampleDB.execSQL("CREATE TABLE IF NOT EXISTS pamiec (Id INTEGER PRIMARY KEY AUTOINCREMENT," +
+                    "email VARCHAR, haslo VARCHAR)");
+            sampleDB.execSQL("Delete from pamiec");
+            sampleDB.execSQL("Insert into pamiec (email,haslo) VALUES ('" + login + "','" + hash + "') ");
+            sampleDB.close();
+        }
 
     }
 
     private void nie_pamiec_hasla()
     {
+        password="";
+        hash="";
+        pamiec="";
+        login_pamiec="";
         sampleDB = this.openOrCreateDatabase(SAMPLE_DB_NAME, MODE_PRIVATE,null);
         sampleDB.execSQL("CREATE TABLE IF NOT EXISTS pamiec (Id INTEGER PRIMARY KEY AUTOINCREMENT," +
                 "email VARCHAR, haslo VARCHAR)");
@@ -115,6 +146,9 @@ public class MainActivity extends AppCompatActivity {
     //sprawdzanie czy uzytkownik zapamietał hasło
     private void sprawdz_haslo()
     {
+        password="";
+        login_pamiec="";
+
         sampleDB = this.openOrCreateDatabase(SAMPLE_DB_NAME, MODE_PRIVATE,null);
         sampleDB.execSQL("CREATE TABLE IF NOT EXISTS pamiec (Id INTEGER PRIMARY KEY AUTOINCREMENT,email VARCHAR, haslo VARCHAR)");
         Cursor d =sampleDB.rawQuery("Select * from pamiec",null);
@@ -147,9 +181,11 @@ public class MainActivity extends AppCompatActivity {
         login_act = email.getText().toString();
         if(login_act.equals(login)) {
             message = "Witaj "+login_act+",\n" +
-                    " Otrzymałeś ten email ponieważ skorzystałeś z opcji przypomnienia hasła. \n" +
-                    " Twoje hasło to: "+zapis_hasla+" \n"+
-                    " Pozdrawiam Zespół TrustCar \n";
+                    " \n "+
+                    "Otrzymałeś ten email ponieważ skorzystałeś z opcji przypomnienia hasła. \n" +
+                    "Twoje hasło to: "+password+"  \n"+
+                    " \n "+
+                    "Pozdrawiam Zespół TrustCar. \n";
             subject = "Przypomnienie hasła TrustCar";
             //Creating SendMail object
 
@@ -162,6 +198,17 @@ public class MainActivity extends AppCompatActivity {
                 Log.i("MainActivity", "" + e);
             }
         }
+    }
+
+    public boolean activeNetwork () {
+        ConnectivityManager cm = (ConnectivityManager)getSystemService(Context.CONNECTIVITY_SERVICE);
+
+        NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
+        boolean isConnected = activeNetwork != null &&
+                activeNetwork.isConnected();
+
+        return isConnected;
+
     }
 
 
@@ -189,33 +236,36 @@ public class MainActivity extends AppCompatActivity {
             public void onClick(View view) {
 
 
-
                 login_act = email.getText().toString();
                 password_act = haslo.getText().toString();
 
                 login_act = login_act.replace(" ","");
-                password_act = password_act.replace(" ","");
+
 
                 SelectDataUser();
-                if(pamiec != null) {
                     hash();
-                }
 
                     if (login_act.equals(login) & login_act.contains("@")) {
-                        if (hash.equals(password) || (password_act.equals(password))) {
+                        Log.i("password_1",password);
+                        Log.i("pasword_asct_1",password_act);
+                        if (hash.equals(password) || (password.equals(password_act))) {
 
                                 //sprawdzanie czy mam zapamiętać hasło
                                 if(checkBox.isChecked())
                                 {
                                     pamiec_hasla();
+
                                 }
                                 Intent c = new Intent(MainActivity.this,MainMenu.class);
                                 c.putExtra("email",login);
                                 startActivity(c);
                                 finish();
+                                //Log.i("password",password);
+                               // Log.i("pasword_asct",password_act);
+                                //Log.i("hash",hash);
                                 }else
                         {
-                            showToast("Nie poprawne hasło");
+                            showToast("Nie poprawne hasło" + password);
                         }
 
                         }else
@@ -242,19 +292,24 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
 
-
-
                 login_act = email.getText().toString();
 
-                if (login_act.contains("@"))
+                SelectDataUser();
+
+                if(login_act.contains("@"))
                 {
-                if(!login_act.equals(""))
+                if(login_act.equals(login))
                 {
-                    sendemail_execiut();
+                    if(activeNetwork()) {
+                        sendemail_execiut();
+                    }else
+                    {
+                        showToast("Brak dostępu do internetu");
+                    }
 
                 }else
                 {
-                    showToast("Wypełnij pole z emailem");
+                    showToast("Brak adresu email w bazie");
                 }
                 }else
                 {
@@ -281,7 +336,7 @@ public class MainActivity extends AppCompatActivity {
                 login_act = email.getText().toString();
                 if (isChecked & login_act!=null)
                 {
-                    pamiec_hasla();
+                    //pamiec_hasla();
                 }
                 else
                 {
